Move frame listener to its own class
