+++ src/main/java/rsfost/loadtime/LoadTimePlugin.java
@@ -26,30 +26,32 @@ package rsfost.loadtime;
 
 import com.google.inject.Provides;
 import javax.inject.Inject;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.Value;
+import lombok.experimental.Accessors;
 import lombok.extern.slf4j.Slf4j;
-import net.runelite.api.*;
+import net.runelite.api.ChatMessageType;
+import net.runelite.api.Client;
+import net.runelite.api.Constants;
 import net.runelite.api.coords.WorldPoint;
-import net.runelite.api.events.ClientTick;
-import net.runelite.api.events.GameStateChanged;
 import net.runelite.api.events.GameTick;
-import net.runelite.api.events.PreMapLoad;
-import net.runelite.client.callback.ClientThread;
 import net.runelite.client.chat.ChatMessageBuilder;
 import net.runelite.client.chat.ChatMessageManager;
 import net.runelite.client.chat.QueuedMessage;
 import net.runelite.client.config.ConfigManager;
+import net.runelite.client.eventbus.EventBus;
 import net.runelite.client.eventbus.Subscribe;
 import net.runelite.client.events.ConfigChanged;
 import net.runelite.client.plugins.Plugin;
 import net.runelite.client.plugins.PluginDescriptor;
 import net.runelite.client.ui.DrawManager;
 import net.runelite.client.util.Text;
-
 import java.awt.Color;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Objects;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
 @Slf4j
@@ -62,13 +64,15 @@ public class LoadTimePlugin extends Plugin
 	@Inject
 	private Client client;
 	@Inject
+	private FrameListener frameListener;
+	@Inject
+	private EventBus eventBus;
+	@Inject
 	private DrawManager drawManager;
 	@Inject
 	private LoadTimeConfig config;
 	@Inject
 	private ChatMessageManager chatMessageManager;
-	@Inject
-	private ClientThread clientThread;
 
 	private Collection<Integer> regions;
 
@@ -76,11 +80,10 @@ public class LoadTimePlugin extends Plugin
 	private long lastGameTickTime;
 	private int lastRegionId;
 
-	private Thread mapLoader;
-	private Scene lastScene;
-	private long mapLoadStartTime;
+	@Getter(AccessLevel.PACKAGE)
+	@Setter(AccessLevel.PACKAGE)
+	@Accessors(fluent = true)
 	private boolean shouldAnnounce;
-	private int mapLoadStartTick;
 
 	@Subscribe
 	public void onGameTick(GameTick tick)
@@ -102,50 +105,6 @@ public class LoadTimePlugin extends Plugin
 		lastRegionId = currentRegionId;
 	}
 
-	public void onFrame()
-	{
-		final long currentTime = System.currentTimeMillis();
-		Player player = client.getLocalPlayer();
-		if (player == null)
-		{
-			lastScene = null;
-			return;
-		}
-
-		Scene scene = player.getWorldView().getScene();
-		if (lastScene != null && lastScene != scene)
-		{
-			final long loadTime = currentTime - mapLoadStartTime;
-			final int startTick = mapLoadStartTick;
-
-			clientThread.invokeLater(() -> {
-				if (!shouldAnnounce && client.getTickCount() == startTick)
-				{
-					return false;
-				}
-				if (shouldAnnounce)
-				{
-					announceLoadTime(loadTime, 2);
-					shouldAnnounce = false;
-				}
-				return true;
-			});
-			mapLoadStartTime = -1;
-		}
-		else if (mapLoader != null && mapLoader.getState() == Thread.State.RUNNABLE && mapLoadStartTime < 0)
-		{
-			mapLoadStartTick = client.getTickCount();
-			mapLoadStartTime = System.currentTimeMillis();
-		}
-		lastScene = scene;
-	}
-
-	@Subscribe
-	public void onPreMapLoad(PreMapLoad event)
-	{
-		mapLoader = Thread.currentThread();
-	}
-
 	@Subscribe
 	public void onConfigChanged(ConfigChanged event)
 	{
@@ -160,13 +119,15 @@ public class LoadTimePlugin extends Plugin
 	protected void startUp() throws Exception
 	{
 		parseRegionIds();
-		drawManager.registerEveryFrameListener(this::onFrame);
+		drawManager.registerEveryFrameListener(frameListener);
+		eventBus.register(frameListener);
 	}
 
 	@Override
 	protected void shutDown() throws Exception
 	{
-		drawManager.unregisterEveryFrameListener(this::onFrame);
+		drawManager.unregisterEveryFrameListener(frameListener);
+		eventBus.unregister(frameListener);
 	}
 
 	@Provides
@@ -196,13 +157,8 @@ public class LoadTimePlugin extends Plugin
 		return true;
 	}
 
-	private void announceLoadTime(long time, int method)
+	void announceLoadTime(long time, int method)
 	{
-		if (!shouldAnnounce)
-		{
-			return;
-		}
-
 		final Color color;
 		if (time < config.fastLoadTime())
 		{
